+++ api/internal/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID         int    `json:"id" db:"id"`
+	ID         uint8  `json:"id" db:"id"`
 	Username   string `json:"username" db:"username"`
 	Password   string `json:"password" db:"password"`
 	FirstName  string `json:"first_name" db:"first_name"`
